+++ rust/Cargo.toml
@@ -6,7 +6,7 @@
 name = "tor"
 version = "0.1.0"
 authors = ["Igor Cota <igor@foundationdevices.com>"]
-edition = "2018"
+edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "staticlib"]
@@ -14,10 +14,10 @@ crate-type = ["cdylib", "staticlib"]
 [dependencies]
 lazy_static = "1.4"
 tokio = { version = "1", features = ["full"] }
-arti-client = { version = "0.12.0", features = ["static"] }
-arti = { version = "1.1.11", features = ["experimental-api", "static"] }
-tor-rtcompat = { version = "0.9.6", features = ["static"] }
-tor-config = "0.9.7"
+arti-client = { version = "0.13.0", features = ["static"] }
+arti = { version = "1.1.12", features = ["experimental-api", "static"] }
+tor-rtcompat = { version = "0.9.7", features = ["static"] }
+tor-config = "0.10.0"
 log = "0.4.20"
 #android_log-sys = "0.3.1"
 rlimit = "0.10.1"
