--- rust/src/error.rs.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2023 Foundation Devices Inc.
-
-SPDX-License-Identifier: GPL-3.0-or-later
